Module Serapi.Serapi_pp

This module includes all of sertop custom Format-based printers for Coq datatypes.

We may want to split it into a library at some point and replace parts of Coq printing/

type 'a pp = Stdlib.Format.formatter -> 'a -> unit

val pp_str : string pp

val pp_opt : 'a pp -> 'a option pp

val pp_list : ?sep:string -> 'a pp -> 'a list pp

val pp_stateid : Stateid.t pp

val pp_feedback : Feedback.feedback pp

val pp_xml : Xml_datatype.xml pp